+++ src/Workspaces/Core/Portable/Workspace/Solution/StateChecksums.cs
@@ -16,89 +16,6 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Serialization;
 
-[DataContract]
-internal readonly struct AssetHint
-{
-    public static readonly AssetHint None = default;
-
-    [DataMember(Order = 0)]
-    public readonly ProjectId? ProjectId;
-    [DataMember(Order = 1)]
-    public readonly DocumentId? DocumentId;
-
-    private AssetHint(ProjectId? projectId, DocumentId? documentId)
-    {
-        ProjectId = projectId;
-        DocumentId = documentId;
-    }
-
-    public static implicit operator AssetHint(ProjectId projectId) => new(projectId, documentId: null);
-    public static implicit operator AssetHint(DocumentId documentId) => new(documentId.ProjectId, documentId);
-}
-
-internal readonly struct ChecksumsAndIds<TId> where TId : IObjectWritable
-{
-    public readonly ChecksumCollection Checksums;
-    public readonly ImmutableArray<TId> Ids;
-
-    private static readonly Func<ObjectReader, TId> s_readId;
-
-    static ChecksumsAndIds()
-    {
-        if (typeof(TId) == typeof(ProjectId))
-        {
-            s_readId = reader => (TId)(object)ProjectId.ReadFrom(reader);
-        }
-        else if (typeof(TId) == typeof(DocumentId))
-        {
-            s_readId = reader => (TId)(object)DocumentId.ReadFrom(reader);
-        }
-        else
-        {
-            throw ExceptionUtilities.Unreachable();
-        }
-    }
-
-    public ChecksumsAndIds(ChecksumCollection checksums, ImmutableArray<TId> ids)
-    {
-        Contract.ThrowIfTrue(ids.Length != checksums.Children.Length);
-
-        Checksums = checksums;
-        Ids = ids;
-    }
-
-    public int Length => Ids.Length;
-    public Checksum Checksum => Checksums.Checksum;
-
-    public void WriteTo(ObjectWriter writer)
-    {
-        this.Checksums.WriteTo(writer);
-        writer.WriteArray(this.Ids, static (writer, value) => value.WriteTo(writer));
-    }
-
-    public static ChecksumsAndIds<TId> ReadFrom(ObjectReader reader)
-    {
-        return new(
-            ChecksumCollection.ReadFrom(reader),
-            reader.ReadArray(s_readId));
-    }
-
-    public Enumerator GetEnumerator()
-        => new(this);
-
-    public struct Enumerator(ChecksumsAndIds<TId> checksumsAndIds)
-    {
-        private readonly ChecksumsAndIds<TId> _checksumsAndIds = checksumsAndIds;
-        private int _index = -1;
-
-        public bool MoveNext()
-            => ++_index < _checksumsAndIds.Length;
-
-        public (Checksum checksum, TId id) Current
-            => (_checksumsAndIds.Checksums.Children[_index], _checksumsAndIds.Ids[_index]);
-    }
-}
-
 internal sealed class SolutionStateChecksums(
     Checksum attributes,
     ChecksumsAndIds<ProjectId> projects,
